Rename: 페이지에 해당하는 파일 컴포넌트에서 각 페이지 폴더로 이동
